+++ catalyst/core/callbacks/__init__.py
@@ -2,8 +2,10 @@
 
 from .checkpoint import CheckpointCallback, IterationCheckpointCallback
 from .criterion import CriterionCallback
-from .logging import ConsoleLogger, TensorboardLogger, VerboseLogger, MetricsManagerCallback, TimerCallback, ValidationManagerCallback
+from .logging import ConsoleLogger, TensorboardLogger, VerboseLogger, MetricsManagerCallback
 from .optimizer import OptimizerCallback
 from .phase import PhaseManagerCallback
 from .scheduler import LRUpdater, SchedulerCallback
+from .timer import TimerCallback
+from .validation import ValidationManagerCallback
 from .wrappers import PhaseBatchWrapperCallback, PhaseWrapperCallback
